@@ -1,24 +1,28 @@
 # ---- All project specifications ---- #
 [project]
 name = "torchtitan"
-description = "A native-PyTorch library for large scale LLM training"
+description = "A PyTorch native library for large-scale model training"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 license = {file = "LICENSE"}
 authors = [
     { name = "PyTorch Team", email = "packages@pytorch.org" },
 ]
 keywords = ["pytorch", "training", "llm"]
 dependencies = [
+    # Stateful Dataloader
+    "torchdata>=0.8.0",
+
     # Hugging Face integrations
-    "datasets>=2.19.0",
+    "datasets>=2.21.0",
 
     # Tokenization
     "blobfile",
     "tiktoken",
 
     # Miscellaneous
-    "tomli>=1.1.0"
+    "tomli>=1.1.0",
+    "fsspec"
 ]
 dynamic = ["version"]
 
@@ -47,9 +51,6 @@ build-backend = "setuptools.build_meta"
 [tool.setuptools.packages.find]
 where = [""]
 include = ["torchtitan*"]
-
-[tool.setuptools.package-data]
-recipes = ["train_configs/*.toml"]
 
 [tool.pytest.ini_options]
 addopts = ["--showlocals"]  # show local variables in tracebacks
