@@ -9,19 +9,20 @@ import gc
 import os
 
 import torch
+
 from torch._guards import active_fake_mode
 from torch._subclasses.fake_tensor import FakeTensorMode
 from torch.distributed._tools.fsdp2_mem_tracker import FSDPMemTracker
 from torch.testing._internal.distributed.fake_pg import FakeStore
 
-from torchtitan import utils
-from torchtitan.config_manager import JobConfig
-from torchtitan.datasets import build_tokenizer
-from torchtitan.float8 import Float8Handler
-from torchtitan.logging import init_logger, logger
-from torchtitan.models import model_name_to_cls, model_name_to_tokenizer, models_config
-from torchtitan.optimizer import build_lr_schedulers, build_optimizers
-from torchtitan.parallelisms import models_parallelize_fns, ParallelDims
+from torchtitan.components.ft import init_ft_manager
+from torchtitan.components.lr_scheduler import build_lr_schedulers
+from torchtitan.components.optimizer import build_optimizers
+from torchtitan.config_manager import ConfigManager, JobConfig
+from torchtitan.distributed import ParallelDims, utils as dist_utils
+from torchtitan.protocols.model_converter import build_model_converters
+from torchtitan.protocols.train_spec import get_train_spec
+from torchtitan.tools.logging import init_logger, logger
 
 
 def estimate_memory(job_config: JobConfig):
@@ -33,34 +34,20 @@ def estimate_memory(job_config: JobConfig):
     # Get the world size
     world_size = int(os.environ["WORLD_SIZE"])
 
-    # fake tensor doesn't work with fused rmsnorm
-    if (
-        job_config.model.norm_type == "fused_rmsnorm"
-        and not job_config.memory_estimation.disable_fake_mode
-    ):
-        logger.info(
-            "Fused RMSNorm is not supported yet under fake estimation mode. "
-            "Switching to rmsnorm."
-        )
-        job_config.model.norm_type = "rmsnorm"
-
-    if job_config.model.norm_type == "compiled_rmsnorm":
-        logger.info("Compiled RMSNorm is not supported yet. Switching to RMSNorm.")
-        job_config.model.norm_type = "rmsnorm"
-
-    if job_config.training.compile or job_config.experimental.enable_compiled_autograd:
+    if job_config.training.compile or job_config.parallelism.enable_compiled_autograd:
         logger.info("Compile mode is not supported yet. Switching to eager mode.")
         job_config.training.compile = False
-        job_config.experimental.enable_compiled_autograd = False
+        job_config.parallelism.enable_compiled_autograd = False
 
+    parallelism_config = job_config.parallelism
     parallel_dims = ParallelDims(
-        dp_shard=job_config.training.data_parallel_shard_degree,
-        dp_replicate=job_config.training.data_parallel_replicate_degree,
-        cp=job_config.experimental.context_parallel_degree,
-        tp=job_config.training.tensor_parallel_degree,
-        pp=job_config.experimental.pipeline_parallel_degree,
+        dp_shard=parallelism_config.data_parallel_shard_degree,
+        dp_replicate=parallelism_config.data_parallel_replicate_degree,
+        cp=parallelism_config.context_parallel_degree,
+        tp=parallelism_config.tensor_parallel_degree,
+        pp=parallelism_config.pipeline_parallel_degree,
         world_size=world_size,
-        enable_loss_parallel=not job_config.training.disable_loss_parallel,
+        enable_loss_parallel=not parallelism_config.disable_loss_parallel,
     )
 
     # only FSDP and HSDP are supported
@@ -85,52 +72,41 @@ def estimate_memory(job_config: JobConfig):
         "fake", rank=int(os.environ["LOCAL_RANK"]), world_size=world_size, store=store
     )
 
+    train_spec = get_train_spec(job_config.model.name)
+
     # build meshes
     world_mesh = parallel_dims.build_mesh(device_type="cuda")
 
-    model_name = job_config.model.name
-
     # build tokenizer
-    tokenizer_type = model_name_to_tokenizer[model_name]
-    tokenizer = build_tokenizer(tokenizer_type, job_config.model.tokenizer_path)
+    tokenizer = train_spec.build_tokenizer_fn(job_config)
 
-    train_context = utils.get_train_context(
+    train_context = dist_utils.get_train_context(
         parallel_dims.loss_parallel_enabled,
-        job_config.experimental.enable_compiled_autograd,
+        job_config.parallelism.enable_compiled_autograd,
     )
 
-    # loss fn can be shared by pipeline-parallel or non-pp execution
-    def loss_fn(pred, labels):
-        return torch.nn.functional.cross_entropy(
-            pred.flatten(0, 1).float(), labels.flatten(0, 1)
-        )
-
     # build model (using meta init)
-    model_cls = model_name_to_cls[model_name]
-    model_config = models_config[model_name][job_config.model.flavor]
-    # set the model configs from training inputs:
-    # 1. norm type to decide which norm layer to use
-    # 2. vocab size from tokenizer
-    # 3. max_seq_len base on inputs
-    model_config.norm_type = job_config.model.norm_type
-    model_config.vocab_size = tokenizer.n_words
-    model_config.max_seq_len = job_config.training.seq_len
+    model_cls = train_spec.cls
+    model_args = train_spec.config[job_config.model.flavor]
+    model_args.update_from_config(job_config, tokenizer)
 
-    with FakeTensorMode() if not job_config.memory_estimation.disable_fake_mode else contextlib.nullcontext():
-
+    with (
+        FakeTensorMode()
+        if not job_config.memory_estimation.disable_fake_mode
+        else contextlib.nullcontext()
+    ):
         logger.info(
-            f"Building {model_name} {job_config.model.flavor} with {model_config}"
+            f"Building {train_spec.name} {job_config.model.flavor} with {model_args}"
         )
         with torch.device("meta"):
-            model = model_cls.from_model_args(model_config)
+            model = model_cls.from_model_args(model_args)
 
-        # a no-op hander if float8 is not enabled
-        float8_handler = Float8Handler(job_config, parallel_dims)
-        # swap to Float8Linear based on float8 configs
-        float8_handler.convert_to_float8_training(model)
+        # Build the collection of model converters. No-op if `model.converters` empty
+        model_converters = build_model_converters(job_config, parallel_dims)
+        model_converters.convert(model)
 
         # apply PT-D DP/TP parallelisms and activation checkpointing
-        models_parallelize_fns[model_name](model, world_mesh, parallel_dims, job_config)
+        train_spec.parallelize_fn(model, world_mesh, parallel_dims, job_config)
 
         model.to_empty(device="cuda")
         if not active_fake_mode():
@@ -138,28 +114,36 @@ def estimate_memory(job_config: JobConfig):
         model.train()
 
         # build optimizer after applying parallelisms to the model
-        optimizers = build_optimizers([model], job_config)
+        ft_manager = init_ft_manager(job_config)
+        optimizers = build_optimizers([model], job_config, ft_manager)
         lr_schedulers = build_lr_schedulers(optimizers.optimizers, job_config)
+        # Post optimizer step model converters hook.
+        # e.g. calculate float8 dynamic amax/scale for all-parameter for FSDP2
+        # where it issues a single all-reduce for all parameters at once for better performance
+        optimizers.register_step_post_hook(
+            lambda *args, **kwargs: model_converters.post_optimizer_hook(model)
+        )
 
-        logger.info(f"Vocab size: {model_config.vocab_size}")
+        logger.info(f"Vocab size: {model_args.vocab_size}")
         # Create a dummy batch instead of loading from a dataset
         batch = (
             torch.randint(
                 0,
-                model_config.vocab_size,
-                (job_config.training.batch_size, model_config.max_seq_len),
+                model_args.vocab_size,
+                (job_config.training.batch_size, model_args.max_seq_len),
                 device="cuda",
             ),
             torch.randint(
                 0,
-                model_config.vocab_size,
-                (job_config.training.batch_size, model_config.max_seq_len),
+                model_args.vocab_size,
+                (job_config.training.batch_size, model_args.max_seq_len),
                 device="cuda",
             ),
         )
         fsdp_memtracker = FSDPMemTracker(mod=model, optm=optimizers.optimizers[0])
         fsdp_memtracker.track_inputs(batch)
 
+        loss_fn = train_spec.build_loss_fn(job_config)
         with fsdp_memtracker:
             for iter_idx in range(2):
                 input_ids, labels = batch
@@ -174,14 +158,10 @@ def estimate_memory(job_config: JobConfig):
                 torch.nn.utils.clip_grad_norm_(
                     model.parameters(), job_config.training.max_norm, foreach=True
                 )
-                # sync float8 amaxes and scales
-                float8_handler.sync_float8_amax_and_scale_history(model)
                 # optimizer step
                 optimizers.step()
                 lr_schedulers.step()
-                # calculate float8 dynamic amax/scale for all-parameter for FSDP2
-                # it issues a single all-reduce for all parameters at once for better performance
-                float8_handler.precompute_float8_dynamic_scale_for_fsdp(model)
+
                 optimizers.zero_grad()
                 print(f"Peak Memory at iter: {iter_idx}")
                 fsdp_memtracker.display_snapshot("peak", units="MiB", tabulate=True)
@@ -205,13 +185,13 @@ def estimate_memory(job_config: JobConfig):
         )
         print(f"Tracker Max: {tracker_peak / gib} GiB")
         if job_config.memory_estimation.disable_fake_mode and peak_active > 0:
-            print(f"Tracker Accuracy: {tracker_peak/peak_active}")
+            print(f"Tracker Accuracy: {tracker_peak / peak_active}")
         gc.enable()
 
 
 if __name__ == "__main__":
-    config = JobConfig()
-    config.parse_args()
+    config_manager = ConfigManager()
+    config = config_manager.parse_args()
     try:
         estimate_memory(config)
     finally:
