@@ -78,7 +78,7 @@ def build_test_list():
         OverrideDefinitions(
             [
                 [
-                    "--training.tensor_parallel_degree 2",
+                    "--parallelism.tensor_parallel_degree 2",
                 ],
             ],
             "2D eager",
@@ -88,22 +88,24 @@ def build_test_list():
             [
                 [
                     "--training.compile",
-                    "--training.tensor_parallel_degree 2",
+                    "--parallelism.tensor_parallel_degree 2",
                 ],
             ],
             "2D compile",
             "2d_compile",
         ),
-        OverrideDefinitions(
-            [
-                [
-                    "--training.tensor_parallel_degree 2",
-                    "--model.norm_type=fused_rmsnorm",
-                ],
-            ],
-            "2D eager with fused_rmsnorm",
-            "2d_eager_fused_rmsnorm",
-        ),
+        # TODO: re-enable this test once the async TP issue is fixed
+        # OverrideDefinitions(
+        #     [
+        #         [
+        #             "--training.compile",
+        #             "--parallelism.tensor_parallel_degree 2",
+        #             "--parallelism.enable_async_tensor_parallel",
+        #         ],
+        #     ],
+        #     "2D async TP compile",
+        #     "2d_asynctp_compile",
+        # ),
         OverrideDefinitions(
             [
                 [
@@ -138,23 +140,35 @@ def build_test_list():
             "Checkpoint Integration Test - Save Model Weights Only bf16",
             "model_weights_only_bf16",
         ),
+        # TODO: enable the following tests once they are fixed
+        # OverrideDefinitions(
+        #     [
+        #         [
+        #             "--parallelism.pipeline_parallel_degree 4",
+        #             "--parallelism.pipeline_parallel_schedule InterleavedZeroBubble",
+        #         ],
+        #     ],
+        #     "PP looped zero bubble test",
+        #     "pp_looped_zero_bubble",
+        #     ngpu=4,
+        # ),
+        # OverrideDefinitions(
+        #     [
+        #         [
+        #             "--parallelism.pipeline_parallel_degree 2",
+        #             "--parallelism.pipeline_parallel_schedule ZBVZeroBubble",
+        #         ],
+        #     ],
+        #     "PP zero bubble test (v shaped)",
+        #     "pp_zbv",
+        #     ngpu=2,
+        # ),
         OverrideDefinitions(
             [
                 [
-                    "--experimental.pipeline_parallel_degree 4",
-                    "--experimental.pipeline_parallel_schedule InterleavedZeroBubble",
-                ],
-            ],
-            "PP looped zero bubble test",
-            "pp_looped_zero_bubble",
-            ngpu=4,
-        ),
-        OverrideDefinitions(
-            [
-                [
-                    "--experimental.pipeline_parallel_degree 2",
-                    "--experimental.pipeline_parallel_schedule 1F1B",
-                    "--training.data_parallel_shard_degree 1",
+                    "--parallelism.pipeline_parallel_degree 2",
+                    "--parallelism.pipeline_parallel_schedule 1F1B",
+                    "--parallelism.data_parallel_shard_degree 1",
                 ],
             ],
             "PP 1D test 1F1B",
@@ -164,9 +178,9 @@ def build_test_list():
         OverrideDefinitions(
             [
                 [
-                    "--experimental.pipeline_parallel_degree 2",
-                    "--experimental.pipeline_parallel_schedule GPipe",
-                    "--training.data_parallel_shard_degree 1",
+                    "--parallelism.pipeline_parallel_degree 2",
+                    "--parallelism.pipeline_parallel_schedule GPipe",
+                    "--parallelism.data_parallel_shard_degree 1",
                 ],
             ],
             "PP 1D test GPipe",
@@ -176,9 +190,15 @@ def build_test_list():
         OverrideDefinitions(
             [
                 [
-                    "--experimental.pipeline_parallel_degree 2",
-                    "--experimental.pipeline_parallel_schedule 1F1B",
-                    "--training.data_parallel_shard_degree 2",
+                    "--parallelism.pipeline_parallel_degree 2",
+                    "--parallelism.pipeline_parallel_schedule 1F1B",
+                    "--parallelism.data_parallel_shard_degree 2",
+                ],
+                [
+                    "--parallelism.pipeline_parallel_degree 2",
+                    "--parallelism.pipeline_parallel_schedule 1F1B",
+                    "--parallelism.pipeline_parallel_layers_per_stage 4",
+                    "--parallelism.data_parallel_shard_degree 2",
                 ],
             ],
             "PP+DP 1F1B 2D test",
@@ -187,9 +207,9 @@ def build_test_list():
         OverrideDefinitions(
             [
                 [
-                    "--experimental.pipeline_parallel_degree 2",
-                    "--experimental.pipeline_parallel_schedule GPipe",
-                    "--training.data_parallel_shard_degree 2",
+                    "--parallelism.pipeline_parallel_degree 2",
+                    "--parallelism.pipeline_parallel_schedule GPipe",
+                    "--parallelism.data_parallel_shard_degree 2",
                 ],
             ],
             "PP+DP GPipe 2D test",
@@ -198,8 +218,8 @@ def build_test_list():
         OverrideDefinitions(
             [
                 [
-                    "--experimental.pipeline_parallel_degree 2",
-                    "--training.tensor_parallel_degree 2",
+                    "--parallelism.pipeline_parallel_degree 2",
+                    "--parallelism.tensor_parallel_degree 2",
                 ],
             ],
             "PP+TP 2D test",
@@ -209,16 +229,16 @@ def build_test_list():
             [
                 [
                     "--checkpoint.enable_checkpoint",
-                    "--experimental.pipeline_parallel_degree 2",
-                    "--training.data_parallel_shard_degree 2",
-                    "--training.tensor_parallel_degree 2",
+                    "--parallelism.pipeline_parallel_degree 2",
+                    "--parallelism.data_parallel_shard_degree 2",
+                    "--parallelism.tensor_parallel_degree 2",
                 ],
                 [
                     "--training.steps 20",
                     "--checkpoint.enable_checkpoint",
-                    "--experimental.pipeline_parallel_degree 2",
-                    "--training.data_parallel_shard_degree 2",
-                    "--training.tensor_parallel_degree 2",
+                    "--parallelism.pipeline_parallel_degree 2",
+                    "--parallelism.data_parallel_shard_degree 2",
+                    "--parallelism.tensor_parallel_degree 2",
                 ],
             ],
             "PP+DP+TP 3D test with save/load resume ckpt",
@@ -228,9 +248,9 @@ def build_test_list():
         OverrideDefinitions(
             [
                 [
-                    "--experimental.pipeline_parallel_degree 2",
-                    "--training.data_parallel_shard_degree 2",
-                    "--training.tensor_parallel_degree 2",
+                    "--parallelism.pipeline_parallel_degree 2",
+                    "--parallelism.data_parallel_shard_degree 2",
+                    "--parallelism.tensor_parallel_degree 2",
                     "--training.compile",
                 ],
             ],
@@ -241,41 +261,46 @@ def build_test_list():
         OverrideDefinitions(
             [
                 [
-                    "--experimental.pipeline_parallel_degree 4",
-                    "--experimental.pipeline_parallel_schedule Interleaved1F1B",
+                    "--parallelism.pipeline_parallel_degree 4",
+                    "--parallelism.pipeline_parallel_schedule Interleaved1F1B",
+                ],
+                [
+                    "--parallelism.pipeline_parallel_degree 4",
+                    "--parallelism.pipeline_parallel_schedule Interleaved1F1B",
+                    "--parallelism.pipeline_parallel_layers_per_stage 1",
                 ],
             ],
             "PP looped 1F1B test",
             "pp_looped_1f1b",
             ngpu=4,
         ),
+        # OverrideDefinitions(
+        #     [
+        #         [
+        #             "--parallelism.pipeline_parallel_degree 2",
+        #             "--parallelism.pipeline_parallel_schedule PipelineScheduleMulti",
+        #             "--parallelism.pipeline_parallel_schedule_csv ./tests/assets/custom_schedule.csv",
+        #         ],
+        #     ],
+        #     "PP with custom pipeline schedule loaded from CSV file",
+        #     "pp_custom_csv",
+        #     ngpu=2,
+        # ),
         OverrideDefinitions(
             [
                 [
-                    "--experimental.pipeline_parallel_degree 2",
-                    "--experimental.pipeline_parallel_schedule PipelineScheduleMulti",
-                    "--experimental.pipeline_parallel_schedule_csv ./tests/assets/custom_schedule.csv",
-                    "--experimental.pipeline_parallel_microbatches 8",
-                ],
+                    "--optimizer.name AdamW --optimizer.implementation foreach",
+                ]
             ],
-            "PP with custom pipeline schedule loaded from CSV file",
-            "pp_custom_csv",
+            "Foreach Optimizer Test",
+            "optimizer_foreach",
             ngpu=2,
         ),
         OverrideDefinitions(
             [
                 [
-                    "--optimizer.name Adam --optimizer.fused",
-                    "--optimizer.name AdamW --optimizer.fused",
-                ]
-            ],
-            "Fused Optimizer Test",
-        ),
-        OverrideDefinitions(
-            [
-                [
-                    "--training.data_parallel_shard_degree=1",
-                    "--training.data_parallel_replicate_degree=4",
+                    "--parallelism.data_parallel_shard_degree=1",
+                    "--parallelism.data_parallel_replicate_degree=4",
                 ]
             ],
             "DDP",
@@ -285,8 +310,8 @@ def build_test_list():
         OverrideDefinitions(
             [
                 [
-                    "--training.data_parallel_shard_degree=2",
-                    "--training.data_parallel_replicate_degree=2",
+                    "--parallelism.data_parallel_shard_degree=2",
+                    "--parallelism.data_parallel_replicate_degree=2",
                 ]
             ],
             "HSDP",
@@ -296,8 +321,20 @@ def build_test_list():
         OverrideDefinitions(
             [
                 [
-                    "--experimental.context_parallel_degree=4",
-                    "--experimental.context_parallel_rotate_method='allgather'",
+                    "--parallelism.data_parallel_shard_degree=4",
+                    "--activation_checkpoint.mode='full'",
+                    "--model.flavor=debugmodel_flex_attn",
+                ]
+            ],
+            "FSDP+FLEX_ATTN",
+            "fsdp+flex_attn",
+            ngpu=4,
+        ),
+        OverrideDefinitions(
+            [
+                [
+                    "--parallelism.context_parallel_degree=4",
+                    "--parallelism.context_parallel_rotate_method='allgather'",
                 ]
             ],
             "CP (allgather)",
@@ -307,8 +344,8 @@ def build_test_list():
         OverrideDefinitions(
             [
                 [
-                    "--experimental.context_parallel_degree=4",
-                    "--experimental.context_parallel_rotate_method='alltoall'",
+                    "--parallelism.context_parallel_degree=4",
+                    "--parallelism.context_parallel_rotate_method='alltoall'",
                 ]
             ],
             "CP (alltoall)",
@@ -318,9 +355,9 @@ def build_test_list():
         OverrideDefinitions(
             [
                 [
-                    "--training.data_parallel_shard_degree=2",
-                    "--training.data_parallel_replicate_degree=2",
-                    "--training.tensor_parallel_degree=2",
+                    "--parallelism.data_parallel_shard_degree=2",
+                    "--parallelism.data_parallel_replicate_degree=2",
+                    "--parallelism.tensor_parallel_degree=2",
                 ]
             ],
             "HSDP+TP",
@@ -330,8 +367,8 @@ def build_test_list():
         OverrideDefinitions(
             [
                 [
-                    "--training.data_parallel_shard_degree=2",
-                    "--experimental.context_parallel_degree=2",
+                    "--parallelism.data_parallel_shard_degree=2",
+                    "--parallelism.context_parallel_degree=2",
                 ]
             ],
             "FSDP+CP",
@@ -341,9 +378,9 @@ def build_test_list():
         OverrideDefinitions(
             [
                 [
-                    "--training.data_parallel_shard_degree=1",
-                    "--training.data_parallel_replicate_degree=2",
-                    "--experimental.context_parallel_degree=2",
+                    "--parallelism.data_parallel_shard_degree=1",
+                    "--parallelism.data_parallel_replicate_degree=2",
+                    "--parallelism.context_parallel_degree=2",
                 ]
             ],
             "HSDP+CP (with dp_shard)",
@@ -353,9 +390,9 @@ def build_test_list():
         OverrideDefinitions(
             [
                 [
-                    "--training.data_parallel_shard_degree=2",
-                    "--training.data_parallel_replicate_degree=2",
-                    "--experimental.context_parallel_degree=2",
+                    "--parallelism.data_parallel_shard_degree=2",
+                    "--parallelism.data_parallel_replicate_degree=2",
+                    "--parallelism.context_parallel_degree=2",
                 ]
             ],
             "HSDP+CP (without dp_shard)",
@@ -365,9 +402,9 @@ def build_test_list():
         OverrideDefinitions(
             [
                 [
-                    "--training.data_parallel_shard_degree=2",
-                    "--training.tensor_parallel_degree=2",
-                    "--experimental.context_parallel_degree=2",
+                    "--parallelism.data_parallel_shard_degree=2",
+                    "--parallelism.tensor_parallel_degree=2",
+                    "--parallelism.context_parallel_degree=2",
                 ]
             ],
             "FSDP+TP+CP",
@@ -378,15 +415,15 @@ def build_test_list():
             [
                 [
                     "--checkpoint.enable_checkpoint",
-                    "--training.tensor_parallel_degree=2",
-                    "--experimental.context_parallel_degree=2",
+                    "--parallelism.tensor_parallel_degree=2",
+                    "--parallelism.context_parallel_degree=2",
                     "--training.enable_cpu_offload",
                     "--optimizer.early_step_in_backward",
                 ],
                 [
-                    "--training.tensor_parallel_degree=2",
-                    "--experimental.context_parallel_degree=2",
-                    "--training.data_parallel_replicate_degree=2",
+                    "--parallelism.tensor_parallel_degree=2",
+                    "--parallelism.context_parallel_degree=2",
+                    "--parallelism.data_parallel_replicate_degree=2",
                     "--training.enable_cpu_offload",
                     "--optimizer.early_step_in_backward",
                 ],
@@ -418,6 +455,34 @@ def build_test_list():
             "test_generate",
             ngpu=2,
         ),
+        OverrideDefinitions(
+            [
+                [
+                    "--parallelism.fsdp_reshard_after_forward always",
+                ],
+            ],
+            "Test always resharding after forward pass",
+            "fsdp_reshard_always",
+            ngpu=2,
+        ),
+        OverrideDefinitions(
+            [
+                [
+                    "--checkpoint.enable_checkpoint",
+                    "--training.steps 10",
+                ],
+                # Save at [dp:4] and load at [dp:2, tp:2]. Note that the dataloader should be
+                # excluded during loading to avoid errors caused by mismatched dp_degree.
+                [
+                    "--checkpoint.enable_checkpoint",
+                    "--checkpoint.exclude_from_loading lr_scheduler,dataloader,optimizer",
+                    "--parallelism.tensor_parallel_degree 2",
+                    "--training.steps 20",
+                ],
+            ],
+            "Optional checkpoint",
+            "optional_checkpoint",
+        ),
     ]
     return integration_tests_flavors
 
@@ -434,7 +499,7 @@ def run_test(test_flavor: OverrideDefinitions, full_path: str, output_dir: str):
     all_ranks = ",".join(map(str, range(test_flavor.ngpu)))
 
     for idx, override_arg in enumerate(test_flavor.override_args):
-        cmd = f"CONFIG_FILE={full_path} NGPU={test_flavor.ngpu} LOG_RANK={all_ranks} ./run_llama_train.sh"
+        cmd = f"CONFIG_FILE={full_path} NGPU={test_flavor.ngpu} LOG_RANK={all_ranks} ./run_train.sh"
         # dump compile trace for debugging purpose
         cmd = f'TORCH_TRACE="{output_dir}/{test_name}/compile_trace" ' + cmd
         if test_name == "fsdp2_memory_estimation":
@@ -492,7 +557,9 @@ def run_tests(args):
 def main():
     parser = argparse.ArgumentParser()
     parser.add_argument("output_dir")
-    parser.add_argument("--config_dir", default="./train_configs")
+    parser.add_argument(
+        "--config_dir", default="./torchtitan/models/llama3/train_configs"
+    )
     parser.add_argument(
         "--test",
         default="all",
